+++ packages/ui/src/form/primitives/field/index.ts
@@ -4,5 +4,16 @@ import { Label } from './label';
 import { Description } from './description';
 import { Details } from './details';
 import { Control } from './control';
+import { VisualLabel } from './visual-label';
+import { VisualDescription } from './visual-description';
 
-export { Root, Item, Label, Description, Details, Control };
+export {
+	Root,
+	Item,
+	Label,
+	Description,
+	Details,
+	Control,
+	VisualLabel,
+	VisualDescription,
+};